% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BT_EM}
\alias{BT_EM}
\title{Fit the Bradley-Terry model using the EM or MM algorithm}
\usage{
BT_EM(W, a, b, maxit = 5000L, epsilon = 0.001)
}
\arguments{
\item{W}{a K*K square matrix of class "dgCMatrix"}

\item{a}{the shape parameter of the gamma prior}

\item{b}{the rate parameter of the gamma prior}

\item{maxit}{the maximum number of iterations}

\item{epsilon}{controls the convergence criteria}
}
\value{
A list containing a K*1 matrix with the pi estimate, the N matrix, the number of iterations, and whether the algorithm converged.
}
\description{
Fit the Bradley-Terry model using the EM or MM algorithm
}

