% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{citations}
\alias{citations}
\title{Statistics Journal Citation Data from Stigler (1994)}
\format{A four by four matrix, where the \eqn{i,j}-th element is the number of times journal \eqn{i} has been cited by journal \eqn{j}.}
\source{
Agresti, A. (2002) \emph{Categorical Data Analysis} (2nd ed.). New York: Wiley
}
\usage{
citations
}
\description{
Extracted from a larger table in Stigler (1994). Inter-journal citation counts for four journals, "Biometrika", "Comm Statist.", "JASA" and "JRSS-B", as used on p448 of Agresti (2002)
}
\details{
In the context of paired comparisons, the 'winner' is the cited journal and the 'loser' is the one doing the citing.
}
\references{
Stigler, S. (1994) Citation patterns in the journals of statistics and probability. \emph{Statistical Science}, \strong{9}, 384-406.
}
\keyword{datasets}

