% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_to_counts.R
\name{code_to_counts}
\alias{code_to_counts}
\title{Convert wins code to counts}
\usage{
code_to_counts(df, code)
}
\arguments{
\item{df}{A three-column data frame. Each row represents a comparison between two items. The first and second columns are the names of the first and second players respectively. The third column gives a code for who won. See examples.}

\item{code}{A numeric vector or character vector, of length two or three (depending on whether there are ties.) The first and second element gives the codes used if the first or second item won respectively. If there are ties, the third element gives the code used in that case. See examples.}
}
\value{
A four-column data frame where the first two columns are the name of the first and second item. The third and fourth column gives the wins count for the first and second item respectively: 1 for a win, 0 for a loss, and 0.5 each for a draw. This data frame is in the correct format to be passed to \code{\link{btdata}}
}
\description{
Convert a three-column data frame in which the third column is a code representing whether the item in column 1 won, lost or (if applicable) drew over/with the item in column 2, to a dataframe with counts (suitable for use in \code{\link{btdata}})
}
\examples{
first <- c("A", "A", "B", "A")
second <- c("B", "B", "C", "C")
df1 <- data.frame(player1 = first, player2 = second, result = c("W1", "W2", "D", "D"))
code_to_counts(df1, c("W1", "W2", "D"))
df2 <- data.frame(item1 = first, item2 = second, code = c(1, 0, 1, .5))
code_to_counts(df2, c(1, 0, .5))
df3 <- data.frame(player1 = first, player2 = second, which_won = c(1,2,2,1))
code_to_counts(df3, c(1,2))

}
