% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_to_matrix.R
\name{graph_to_matrix}
\alias{graph_to_matrix}
\title{Converts a graph representation of wins into a square matrix.}
\usage{
graph_to_matrix(g)
}
\arguments{
\item{g}{A directed \code{\link[igraph]{igraph}} object. If \code{g} is unweighted, an edge from node \eqn{i} to \eqn{j} represents one win for player \eqn{i} over player \eqn{j} (and \eqn{n} edges from \eqn{i} to \eqn{j} represents \eqn{n} wins). If \code{g} is weighted, the weight on an edge from node \eqn{i} to \eqn{j} represents the number of times player \eqn{i} has beaten player \eqn{j}.}
}
\value{
A \eqn{K} by \eqn{K} sparse Matrix, where \eqn{K} is the number of players. The \eqn{i,j}-th element of this matrix is the number of times player \eqn{i} beats player \eqn{j}.
}
\description{
\code{graph_to_matrix} is a data manipulation function that converts a graph representation of wins data into a square matrix.
}
\examples{
g <- igraph::make_graph(c(1,2, 1,2, 2,3, 3,4, 5,6), directed = TRUE)
graph_to_matrix(g)

g2 <- igraph::make_graph(c(1,2, 2,3, 3,4, 5,6), directed = TRUE)
igraph::E(g2)$weight <- 1:4
graph_to_matrix(g2)
}
\seealso{
\code{\link{btfit}}.
}
