% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairs_to_matrix.R
\name{pairs_to_matrix}
\alias{pairs_to_matrix}
\title{Converts a data frame of paired results into a square matrix.}
\usage{
pairs_to_matrix(df)
}
\arguments{
\item{df}{A data frame (either a \code{\link[base]{data.frame}} or an object that inherits that class, such as a \code{\link[dplyr]{tbl_df}} or a \code{\link[data.table]{data.table}}). It must have 3 or 4 columns:
\describe{
  \item{Ordered}{If a 3-column data-frame, the first column contains the name of the winning player, the second column contains the name of the losing player and the third columns contains the number of times that the winner has beaten the loser. Multiple entries for the same pair of teams are handled correctly.}
  \item{Unordered}{If a 4-column data-frame, the first column contains the name of Player 1, the second column contains the name of Player 2, the third column contains the number of times that Player 1 has beaten Player 2 and the fourth column contains the number of times Player 2 has beaten Player 1. Multiple entries for the same pair of teams are handled correctly.}
}
In both cases, if there is a repeated pairing in \code{df}, \code{pairs_to_matrix} aggregates the number of wins, as one would expect.}
}
\value{
A \eqn{K} by \eqn{K} sparse Matrix, where \eqn{K} is the number of players. The \eqn{i,j}-th element of this matrix is the number of times player \eqn{i} beats player \eqn{j}. This output is in the correct format to pass as the \code{W} argument to \code{\link{connected_components}} and \code{\link{btfit}}.
}
\description{
\code{pairs_to_matrix} is a data manipulation function that converts a data frame of paired results (either ordered or unordered) into a square matrix. The output is of the correct form to be passed as the \code{W} argument into the \code{\link{connected_components}} and \code{\link{btfit}} functions in this package.
}
\examples{
winner <- c("A", "A", "A", "A", "A", "B", "B", "B", "C", "C", "E", "F", "F")
loser <- c("B", "B", "C", "D", "F", "D", "A", "F", "F", "G", "G", "A", "C")
df <- data.frame(winner = winner, loser = loser, num_wins = 1:13)
pairs_to_matrix(df)

df2 <- data.frame(team1 = c("A", "A", "B", "C"),
                  team2 = c("B", "C", "C", "D"),
                  wins1 = 1:4, wins2 = 5:8)
pairs_to_matrix(df2)
}
\seealso{
\code{\link{connected_components}}, \code{\link{btfit}}.
}

