% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBT.R
\docType{package}
\name{SBT}
\alias{SBT}
\alias{SBT-package}
\title{A package for fitting the Bradley-Terry model to (potentially) large and sparse data sets.}
\description{
A package for fitting the Bradley-Terry model to (potentially) large and sparse data sets.
}

