% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.btfit.R
\name{coef.btfit}
\alias{coef.btfit}
\title{Extract coefficients of a 'btfit' object}
\usage{
\method{coef}{btfit}(object, subset = NULL, ref = NULL, as_df = TRUE, ...)
}
\arguments{
\item{object}{An object of class "btfit", typically the result \code{ob} of \code{ob <- btfit(..)}. See \code{\link{btfit}}.}

\item{subset}{A character vector of names of components (i.e. a subset of names(object$pi))}

\item{ref}{the reference item, either a string with the item name, 1 or NULL. If NULL, then the coefficients are constrained such that their mean is zero. If an item name is given, they are shifted so that the coefficient for the ref item is zero. If there is more than one component, the components that do not include the ref item will be treated as if ref = NULL. If ref = 1, then the first item of each component is made the reference item.}

\item{as_df}{Logical scalar, determining class of output. If \code{TRUE}, the function returns a data frame. If \code{FALSE} (the default), the function returns a matrix (or list of matrices).}

\item{...}{other arguments}
}
\value{
If as_df = TRUE, a data frame a numeric vector of estimated coefficients, where the first column is the component the item is in, the second column in the item and the third column in the coefficient. If as_df = FALSE, then a numeric vector is returned if the model is fitted on the full dataset, or else a list of numeric vectors is returned, one for each fully connected component. Within each component, the items are arranged by estimate, in descending order.
}
\description{
Extract coefficients of a 'btfit' object
}

