% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_components.R
\name{select_components}
\alias{select_components}
\title{Subset a btdata object}
\usage{
select_components(btdata, subset, return_graph = FALSE)
}
\arguments{
\item{btdata}{A \code{\link{btdata}} object}

\item{subset}{A condition for selecting a subset of the components. This can either be a character vector of names of the components, a single predicate function, or a logical vector of the same length as the number of components (i.e. of length \code{length(btdata$components)}).}

\item{return_graph}{Logical. If TRUE, an igraph object representing the comparison graph will be returned.}
}
\value{
A \code{\link{btdata}} object, which is a list containing:
\item{wins}{A square matrix, where the \eqn{i,j}-th element is the number of times item \eqn{i} has beaten item \eqn{j}.}
\item{components}{A list of the fully-connected components. The names of the list preserve the names of the original \code{btdata} object.}
\item{graph}{The comparison graph of the data (if return_graph = TRUE).}
}
\description{
Subset a btdata object by selecting components from it.
}
\seealso{
\code{\link{btdata}}, \code{\link{btfit}}
}

