% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.btfit.R
\name{simulate_BT}
\alias{simulate_BT}
\alias{simulate.btfit}
\title{Simulate data from a specified Bradley-Terry model}
\usage{
simulate_BT(pi, N, nsim = 1, seed = NULL, result_class = c("sparseMatrix",
  "btdata"))

\method{simulate}{btfit}(object, nsim = 1, seed = NULL,
  result_class = c("sparseMatrix", "btdata"), ...)
}
\arguments{
\item{pi}{a numeric vector, with all values finite and positive.}

\item{N}{a symmetric, numeric matrix with dimensions the same as \code{length(pi)}.  The elements should
be numeric representations of non-negative integers.}

\item{nsim}{a scalar integer, the number of datasets to be generated.}

\item{seed}{an object specifying if and how the random number generator should be initialized (‘seeded’).
For details see \code{\link{simulate}}.}

\item{result_class}{a character vector specifying whether the generated datasets should be of class
"sparseMatrix" or of class "btdata".  The first match among those alternatives is used.}

\item{object}{An object of class "btfit", typically the result \code{ob} of \code{ob <- btfit(..)}. See \code{\link{btfit}}.}

\item{...}{Other arguments}
}
\value{
a list of length \code{nsim} of simulated datasets, each dataset being a sparse matrix with the
same dimensions as \code{N}.
}
\description{
Simulate data from a specified Bradley-Terry model
}
\seealso{
\code{\link{btfit}}
}
\author{
David Firth
}
