% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.btfit.R
\name{simulate_BT}
\alias{simulate.btfit}
\alias{simulate_BT}
\title{Simulate data from a specified Bradley-Terry model}
\usage{
simulate_BT(pi, N, nsim = 1, seed = NULL, result.class = c("sparseMatrix",
  "btdata"))
}
\arguments{
\item{pi}{a numeric vector, with all values finite and positive.}

\item{N}{a symmetric, numeric matrix with dimensions the same as \code{length(pi)}.  The elements should
be numeric representations of non-negative integers.}

\item{nsim}{a scalar integer, the number of datasets to be generated.}

\item{seed}{an object specifying if and how the random number generator should be initialized (‘seeded’).
For details see \code{\link{simulate}}.}

\item{result.class}{a character vector specifying whether the generated datasets should be of class 
"sparseMatrix" or of class "btfit".  The first match among those alternatives is used.}
}
\value{
a list of length \code{nsim} of simulated datasets, each dataset being a sparse matrix with the 
same dimensions as \code{N}.
}
\description{
Simulate data from a specified Bradley-Terry model
}
\examples{
library(BradleyTerry2)
data(citations)
cit.btdata <- btdata(citations)
citmodel <- btfit(cit.btdata, a = 1)
simulate(citmodel, nsim = 3, seed = 1987)
}
\author{
David Firth
}

