% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.btfit.R
\name{summary.btfit}
\alias{summary.btfit}
\title{Summarizing Bradley-Terry Fits}
\usage{
\method{summary}{btfit}(object, subset = NULL, ref = NULL, ...)
}
\arguments{
\item{object}{An object of class "btfit", typically the result \code{ob} of \code{ob <- btfit(..)}. See \code{\link{btfit}}.}

\item{subset}{A character vector of names of components (i.e. a subset of names(object$pi))}

\item{ref}{the reference item, either a string with the item name, 1 or NULL. If NULL, then the coefficients are constrained such that their mean is zero. If an item name is given, they are shifted so that the coefficient for the ref item is zero. If there is more than one component, the components that do not include the ref item will be treated as if ref = NULL. If ref = 1, then the first item of each component is made the reference item.}

\item{...}{other arguments}
}
\value{
An S3 object of class \code{"summary.btfit"}. It is a list containing the following components:
\item{item_summary}{A tibble with columns for the item name, its coefficient, the standard error and the component it is in.}
\item{component_summary}{A tibble with a row for each component in the \code{btfit} object (named according to the original \code{btdata$components}, with the number of items in the component, the number of iterations the fitting algorithm ran for, and whether it converged.}
}
\description{
\code{summary} method for class "btfit"
}

