% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_data}
\alias{toy_data}
\title{A toy data set for the \code{BradleyTerryScalable} pacakge}
\format{A data frame with 13 rows and 3 variables:
\describe{
  \item{player1}{The name of player1}
  \item{player2}{The name of player2}
  \item{wins}{The number of times player1 has beaten player2}
}}
\usage{
toy_data
}
\description{
A toy data set, where the underlying comparison graph of the players is not fully connected.
}
\keyword{datasets}

