% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.btfit.R
\name{vcov.btfit}
\alias{vcov.btfit}
\title{Calculate variance-covariance matrix for a btfit object}
\usage{
\method{vcov}{btfit}(object, subset = NULL, ref = NULL, ...)
}
\arguments{
\item{object}{An object of class "btfit", typically the result \code{ob} of \code{ob <- btfit(..)}. See \code{\link{btfit}}.}

\item{subset}{A character vector of names of components (i.e. a subset of names(object$pi))}

\item{ref}{the reference item, either a string with the item name, 1 or NULL. If NULL, then the coefficients are constrained such that their mean is zero. If an item name is given, they are shifted so that the coefficient for the ref item is zero. If there is more than one component, the components that do not include the ref item will be treated as if ref = NULL. If ref = 1, then the first item of each component is made the reference item.}

\item{...}{other arguments}
}
\value{
A square numeric matrix, which is a non-full-rank variance-covariance matrix for the estimates in \code{coef(object, subset = subset, ref = ref)}; or a list of such matrices if \code{object} has more than one component.
}
\description{
Calculate variance-covariance matrix for a btfit object
}

